segundo avance

+++ TrabajoFinal-Algoritmos/src/FactorialIterativo.java
@@ -1,7 +1,7 @@
 public class FactorialIterativo {
-    long factorial(int n){
+    long factorial(long n){
         long resultado = 1;
-        for (int i = 1; i <= n; i++){
+        for (long i = 1; i <= n; i++){
             resultado = resultado * i;
         }
         return resultado;
